+++ api/src/main/java/shop/ink3/api/common/advice/GlobalExceptionHandler.java
@@ -18,7 +18,7 @@ import shop.ink3.api.payment.exception.PaymentProcessorFailException;
 import shop.ink3.api.user.common.exception.DormantException;
 import shop.ink3.api.user.common.exception.InvalidPasswordException;
 import shop.ink3.api.user.common.exception.WithdrawnException;
-import shop.ink3.api.user.point.exception.PointHistoryAlreadyCanceledException;
+import shop.ink3.api.user.point.history.exception.PointHistoryAlreadyCanceledException;
 import shop.ink3.api.user.user.exception.InsufficientPointException;
 
 @RestControllerAdvice
@@ -85,24 +85,26 @@ public class GlobalExceptionHandler {
     }
 
     @ExceptionHandler(InsufficientBookStockException.class)
-    public ResponseEntity<CommonResponse<String>> handleInsufficientBookStockException(InsufficientBookStockException e) {
+    public ResponseEntity<CommonResponse<String>> handleInsufficientBookStockException(
+            InsufficientBookStockException e) {
         return ResponseEntity
                 .status(HttpStatus.CONFLICT)
-                .body(CommonResponse.error(HttpStatus.CONFLICT, "Book stock insufficient.",e.getMessage()));
+                .body(CommonResponse.error(HttpStatus.CONFLICT, "Book stock insufficient.", e.getMessage()));
     }
 
     @ExceptionHandler(PaymentParserFailException.class)
     public ResponseEntity<CommonResponse<String>> handlePaymentParserFailException(PaymentParserFailException e) {
         return ResponseEntity
                 .status(HttpStatus.BAD_REQUEST)
-                .body(CommonResponse.error(HttpStatus.BAD_REQUEST, "Payment parsing error.",e.getMessage()));
+                .body(CommonResponse.error(HttpStatus.BAD_REQUEST, "Payment parsing error.", e.getMessage()));
     }
 
     @ExceptionHandler(PaymentProcessorFailException.class)
     public ResponseEntity<CommonResponse<String>> handlePaymentProcessorFailException(PaymentProcessorFailException e) {
         return ResponseEntity
                 .status(HttpStatus.BAD_REQUEST)
-                .body(CommonResponse.error(HttpStatus.BAD_REQUEST, "Payment processor Approve fail error.",e.getMessage()));
+                .body(CommonResponse.error(HttpStatus.BAD_REQUEST, "Payment processor Approve fail error.",
+                        e.getMessage()));
     }
 
     @ExceptionHandler(Exception.class)
