+++ api/src/main/java/shop/ink3/api/payment/controller/PaymentController.java
@@ -10,16 +10,11 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import shop.ink3.api.common.dto.CommonResponse;
 import shop.ink3.api.payment.dto.PaymentConfirmRequest;
 import shop.ink3.api.payment.dto.PaymentResponse;
-import shop.ink3.api.payment.entity.Payment;
 import shop.ink3.api.payment.service.PaymentService;
-import shop.ink3.api.user.point.dto.PointHistoryCreateRequest;
-import shop.ink3.api.user.point.entity.PointHistory;
-import shop.ink3.api.user.point.eventListener.PointEventListener;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -52,21 +47,21 @@ public class PaymentController {
     @PostMapping("/{orderId}/cancel")
     public ResponseEntity<CommonResponse<Void>> cancelPayment(
             @PathVariable long orderId,
-            @RequestHeader("X-User-Id") long userId){
-        paymentService.cancelPayment(orderId, userId);  
+            @RequestHeader("X-User-Id") long userId) {
+        paymentService.cancelPayment(orderId, userId);
         return ResponseEntity.noContent().build();
     }
 
     // 결제 결과 조회
     @GetMapping("/{orderId}")
-    public ResponseEntity<CommonResponse<PaymentResponse>> getPayment(@PathVariable long orderId){
+    public ResponseEntity<CommonResponse<PaymentResponse>> getPayment(@PathVariable long orderId) {
         return ResponseEntity
                 .ok(CommonResponse.success(paymentService.getPayment(orderId)));
     }
 
     // 결제 내역 삭제
     @DeleteMapping("/{orderId}")
-    public ResponseEntity<CommonResponse<Void>> deletePayment(@PathVariable long orderId){
+    public ResponseEntity<CommonResponse<Void>> deletePayment(@PathVariable long orderId) {
         paymentService.deletePayment(orderId);
         return ResponseEntity.noContent().build();
     }
