--- api/src/main/java/shop/ink3/api/user/point/entity/PointHistory.java
@@ -1,60 +1,0 @@
-package shop.ink3.api.user.point.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import shop.ink3.api.user.user.entity.User;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
-@Getter
-@Table(name = "point_histories")
-public class PointHistory {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "origin_id")
-    private PointHistory origin;
-
-    @Column(nullable = false)
-    private Integer delta;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private PointHistoryStatus status;
-
-    @Column(nullable = false)
-    private String description;
-
-    @Column(nullable = false)
-    private LocalDateTime createdAt;
-
-    public void update(int delta, PointHistoryStatus status, String description) {
-        this.delta = delta;
-        this.status = status;
-        this.description = description;
-    }
-}
