--- api/src/main/java/shop/ink3/api/user/point/entity/PointHistoryStatus.java
@@ -1,14 +1,0 @@
-package shop.ink3.api.user.point.entity;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum PointHistoryStatus {
-    EARN("적립"),
-    USE("사용"),
-    CANCEL("취소");
-
-    private final String label;
-}
