--- api/src/main/java/shop/ink3/api/user/point/eventListener/PointHistoryAfterCancelPaymentEven.java
@@ -1,7 +1,0 @@
-package shop.ink3.api.user.point.eventListener;
-
-import java.util.List;
-import shop.ink3.api.order.orderPoint.entity.OrderPoint;
-
-public record PointHistoryAfterCancelPaymentEven (long orderId, List<OrderPoint> orderPoints){
-}
