--- api/src/main/java/shop/ink3/api/user/point/eventListener/PointHistoryAfterPaymentEven.java
@@ -1,4 +1,0 @@
-package shop.ink3.api.user.point.eventListener;
-
-public record PointHistoryAfterPaymentEven(long userId ,long orderId, int paymentAmount, int usedPointAmount) {
-}
