--- api/src/main/java/shop/ink3/api/user/point/exception/PointHistoryAlreadyCanceledException.java
@@ -1,7 +1,0 @@
-package shop.ink3.api.user.point.exception;
-
-public class PointHistoryAlreadyCanceledException extends RuntimeException {
-    public PointHistoryAlreadyCanceledException(long pointHistoryId) {
-        super("This point transaction has already been canceled. ID: %d".formatted(pointHistoryId));
-    }
-}
