--- api/src/main/java/shop/ink3/api/user/point/exception/PointHistoryNotFoundException.java
@@ -1,9 +1,0 @@
-package shop.ink3.api.user.point.exception;
-
-import shop.ink3.api.common.exception.NotFoundException;
-
-public class PointHistoryNotFoundException extends NotFoundException {
-    public PointHistoryNotFoundException(long pointHistoryId) {
-        super("PointHistory not found. ID: %d".formatted(pointHistoryId));
-    }
-}
