+++ api/src/main/java/shop/ink3/api/user/point/history/entity/PointHistoryStatus.java
@@ -1,0 +1,14 @@
+package shop.ink3.api.user.point.history.entity;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum PointHistoryStatus {
+    EARN("적립"),
+    USE("사용"),
+    CANCEL("취소");
+
+    private final String label;
+}
