+++ api/src/main/java/shop/ink3/api/user/point/history/eventListener/PointHistoryAfterPaymentEven.java
@@ -1,0 +1,4 @@
+package shop.ink3.api.user.point.history.eventListener;
+
+public record PointHistoryAfterPaymentEven(long userId, long orderId, int paymentAmount, int usedPointAmount) {
+}
