+++ api/src/main/java/shop/ink3/api/user/point/history/exception/PointHistoryNotFoundException.java
@@ -1,0 +1,9 @@
+package shop.ink3.api.user.point.history.exception;
+
+import shop.ink3.api.common.exception.NotFoundException;
+
+public class PointHistoryNotFoundException extends NotFoundException {
+    public PointHistoryNotFoundException(long pointHistoryId) {
+        super("PointHistory not found. ID: %d".formatted(pointHistoryId));
+    }
+}
