+++ api/src/main/java/shop/ink3/api/user/point/policy/exception/PointPolicyNotFoundException.java
@@ -1,0 +1,9 @@
+package shop.ink3.api.user.point.policy.exception;
+
+import shop.ink3.api.common.exception.NotFoundException;
+
+public class PointPolicyNotFoundException extends NotFoundException {
+    public PointPolicyNotFoundException(long pointPolicyId) {
+        super("Point policy not found. ID: " + pointPolicyId);
+    }
+}
