+++ api/src/main/java/shop/ink3/api/user/point/policy/repository/PointPolicyQuerydslRepository.java
@@ -1,0 +1,7 @@
+package shop.ink3.api.user.point.policy.repository;
+
+import shop.ink3.api.user.point.policy.dto.PointPolicyStatisticsResponse;
+
+public interface PointPolicyQuerydslRepository {
+    PointPolicyStatisticsResponse getPointPolicyStatistics();
+}
