+++ api/src/main/java/shop/ink3/api/user/point/policy/repository/PointPolicyRepository.java
@@ -1,0 +1,9 @@
+package shop.ink3.api.user.point.policy.repository;
+
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import shop.ink3.api.user.point.policy.entity.PointPolicy;
+
+public interface PointPolicyRepository extends JpaRepository<PointPolicy, Long>, PointPolicyQuerydslRepository {
+    Optional<PointPolicy> findByIsActive(boolean isActive);
+}
